% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model_EV_cargo_prediction.R
\name{convert_settings_EV_cargo_prediction}
\alias{convert_settings_EV_cargo_prediction}
\title{Convert settings to correct settings format for EV_cargo prediction.}
\usage{
convert_settings_EV_cargo_prediction(settings, all_EV_cargo, validation = TRUE, single = TRUE)
}
\arguments{
\item{settings}{A list of lists. Eeach sublist contains the following elements: .$name: name of the setting; .$from: name(s) of the EV_cargo(s) active in the setting of interest; .$response: the observed target response: indicate for a gene whether it was a target or not in the setting of interest.}

\item{all_EV_cargo}{A character vector of possible EV_cargo that will be considered for the EV_cargo activity state prediction.}

\item{validation}{TRUE if seetings need to be prepared for validation of EV_cargo activity state predictions (this implies that the true active EV_cargo of a setting is known); FALSE for application purposes when the true active EV_cargo(s) is/are not known.}

\item{single}{TRUE if feature importance scores for EV_cargo will be calculated by looking at ligans individually. FALSE if the goal is to calculate the feature importance scores via sophisticated classification algorithms like random forest.}
}
\value{
A list with following elements: $name, $EV_cargo: name of active EV_cargo(s) (only if validation is TRUE), $from (EV_cargo(s) that will be tested for activity prediction), $response
}
\description{
\code{convert_settings_EV_cargo_prediction} Converts settings to correct settings format for EV_cargo activity prediction. In this prediction problem, EV_cargo (out of a set of possibly active EV_cargo) will be ranked based on feature importance scores. The format can be made suited for: 1) validation of EV_cargo activity state prediction by calculating individual feature importane scores or 2) feature importance based on models with embedded feature importance determination; applications in which EV_cargo need to be scores based on their possible upstream activity: 3) by calculating individual feature importane scores or 4) feature importance based on models with embedded feature importance determination.
}
\examples{
\dontrun{
settings = lapply(expression_settings_validation,convert_expression_settings_evaluation)
EV_cargo = unlist(extract_EV_cargo_from_settings(settings,combination = FALSE))
settings_EV_cargo_pred = convert_settings_EV_cargo_prediction(settings, EV_cargo, validation = TRUE, single = TRUE)
}
}
