% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application_visualization.R
\name{diagrammer_format_signaling_graph}
\alias{diagrammer_format_signaling_graph}
\title{Prepare extracted EV_cargo-target signaling network for visualization with DiagrammeR.}
\usage{
diagrammer_format_signaling_graph(signaling_graph_list, EV_cargo_all,targets_all, sig_color = "steelblue", gr_color = "orange")
}
\arguments{
\item{signaling_graph_list}{A list of two elements: sig: a data frame/ tibble containg weighted EV_cargo-receptor and signaling interactions (from, to, weight); and gr: a data frame/tibble containng weighted gene regulatory interactions (from, to, weight)}

\item{EV_cargo_all}{A character vector of one or more EV_cargo of interest}

\item{targets_all}{A character vector of one or more target genes of interest}

\item{sig_color}{The color for EV_cargo-signaling edges and the EV_cargo node. Default: steelblue.}

\item{gr_color}{The color for the gene regulatory edges and the target node. Default: orange.}
}
\value{
A DiagrammeR Graph object ready for visualization via DiagrammeR::render_graph.
}
\description{
\code{diagrammer_format_signaling_graph} Prepare extracted EV_cargo-target signaling network for visualization with DiagrammeR.
}
\examples{
\dontrun{
weighted_networks = construct_weighted_networks(lr_network, sig_network, gr_network,source_weights_df)
EV_cargo = list("TNF","BMP2",c("IL4","IL13"))
EV_cargo_tf_matrix = construct_EV_cargo_tf_matrix(weighted_networks, EV_cargo, ltf_cutoff = 0.99, algorithm = "PPR", damping_factor = 0.5,EV_cargo_as_cols = TRUE)
all_EV_cargo = c("BMP2")
all_targets = c("HEY1")
top_n_regulators = 2
EV_cargo_target_signaling_list = get_EV_cargo_signaling_path(EV_cargo_tf_matrix,all_EV_cargo,all_targets,top_n_regulators,weighted_networks)
graph = diagrammer_format_signaling_graph(EV_cargo_target_signaling_list, all_EV_cargo,all_targets)
# DiagrammeR::render_graph(graph, layout = "tree")
}
}
