% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EV-Net-package.R
\docType{package}
\name{EV-Net-package}
\alias{EV-Net}
\alias{EV-Net-package}
\title{EV-Net: Modification of the Nichenet method for Extracellulas Vesicles-mediated Communication}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/torrejoNia/EV-Net}
  \item Report bugs at \url{https://github.com/torrejoNia/EV-Net/issues}
}

}
\keyword{internal}
