% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model_target_prediction.R
\name{evaluate_target_prediction}
\alias{evaluate_target_prediction}
\title{Evaluation of target gene prediction.}
\usage{
evaluate_target_prediction(setting,EV_cargo_target_matrix, EV_cargo_position = "cols")
}
\arguments{
\item{setting}{A list containing the following elements: .$name: name of the setting; .$from: name(s) of the EV_cargo(s) active in the setting of interest; .$response: named logical vector indicating whether a target is a TRUE target of the possibly active EV_cargo(s) or a FALSE.}

\item{EV_cargo_target_matrix}{A matrix of EV_cargo-target probabilty scores (or discrete target assignments).}

\item{EV_cargo_position}{Indicate whether the EV_cargo in the EV_cargo-target matrix are in the rows ("rows") or columns ("cols"). Default: "cols"}
}
\value{
A data.frame with following variables: setting, EV_cargo nd for probabilistic predictions: auroc, aupr, aupr_corrected (aupr - aupr for random prediction), sensitivity_roc (proxy measure, inferred from ROC), specificity_roc (proxy measure, inferred from ROC), mean_rank_GST_log_pval (-log10 of p-value of mean-rank gene set test), pearson (correlation coefficient), spearman (correlation coefficient); whereas for categorical predictions: accuracy, recall, specificity, precision, F1, F0.5, F2, mcc, informedness, markedness, fisher_pval_log (which is -log10 of p-value fisher exact test), fisher odds.\cr
"mean_rank_GST_log_pval" will only be included in the dataframe if limma is installed. From NicheNet v2.1.7 onwards, limma is no longer a hard dependency of NicheNet.
}
\description{
\code{evaluate_target_prediction} Evaluate how well the model (i.e. the inferred EV_cargo-target probability scores) is able to predict the observed response to a EV_cargo (e.g. the set of DE genes after treatment of cells by a EV_cargo). It shows several classification evaluation metrics for the prediction. Different classification metrics are calculated depending on whether the input EV_cargo-target matrix contains probability scores for targets or discrete target assignments.
}
\examples{
\dontrun{
weighted_networks = construct_weighted_networks(lr_network, sig_network, gr_network, source_weights_df)
setting = lapply(expression_settings_validation[1],convert_expression_settings_evaluation)
EV_cargo = extract_EV_cargo_from_settings(setting)
EV_cargo_target_matrix = construct_EV_cargo_target_matrix(weighted_networks, EV_cargo)
perf1 = lapply(setting,evaluate_target_prediction,EV_cargo_target_matrix)
print(head(perf1))
perf2 = lapply(setting,evaluate_target_prediction,make_discrete_EV_cargo_target_matrix(EV_cargo_target_matrix))
}
}
