% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EV-Net-package.R
\docType{package}
\name{EVNet-package}
\alias{EVNet}
\alias{EVNet-package}
\title{EVNet: Extracellular vesicles-mediated intercellular communication networks}
\description{
This package allow to explore the predicted effect of EVs proteomic or transcriptomic cargo in their recipient tissue of interest.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/torrejoNia/EV-Net}
  \item Report bugs at \url{https://github.com/torrejoNia/EV-Net/issues}
}

}
\keyword{internal}
