% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application_visualization.R
\name{get_EV_cargo_signaling_path}
\alias{get_EV_cargo_signaling_path}
\title{Get EV_cargo-target signaling paths between EV_cargo(s) and target gene(s) of interest}
\usage{
get_EV_cargo_signaling_path(EV_cargo_tf_matrix, EV_cargo_all, targets_all, top_n_regulators = 4, weighted_networks, EV_cargo_position = "cols")
}
\arguments{
\item{EV_cargo_tf_matrix}{A matrix of EV_cargo-regulator probability scores}

\item{EV_cargo_all}{A character vector of one or more EV_cargo of interest}

\item{targets_all}{A character vector of one or more target genes of interest}

\item{top_n_regulators}{The number of top regulators that should be included in the EV_cargo-target signaling network. Top regulators are regulators that score both high for being upstream of the target gene(s) and high for being downstream of the EV_cargo. Default: 4.}

\item{weighted_networks}{A list of two elements: lr_sig: a data frame/ tibble containg weighted EV_cargo-receptor and signaling interactions (from, to, weight); and gr: a data frame/tibble containng weighted gene regulatory interactions (from, to, weight)}

\item{EV_cargo_position}{Indicate whether the EV_cargo in the EV_cargo-target matrix are in the rows ("rows") or columns ("cols"). Default: "cols".}

\item{minmax_scaling}{Indicate whether the weights of both dataframes should be min-max scaled between 0.75 and 1. Default: FALSE.}
}
\value{
A list containing 2 elements (sig and gr): the integrated weighted EV_cargo-signaling and gene regulatory networks data frame / tibble format with columns: from, to, weight
}
\description{
\code{get_EV_cargo_signaling_path} Extract possible signaling paths between a EV_cargo and target gene of interest. The most highly weighted path(s) will be extracted.
}
\examples{
\dontrun{
weighted_networks = construct_weighted_networks(lr_network, sig_network, gr_network,source_weights_df)
EV_cargo = list("TNF","BMP2",c("IL4","IL13"))
EV_cargo_tf_matrix = construct_EV_cargo_tf_matrix(weighted_networks, EV_cargo, ltf_cutoff = 0.99, algorithm = "PPR", damping_factor = 0.5,EV_cargo_as_cols = TRUE)
all_EV_cargo = c("BMP2")
all_targets = c("HEY1")
top_n_regulators = 2
EV_cargo_target_signaling_list = get_EV_cargo_signaling_path(EV_cargo_tf_matrix,all_EV_cargo,all_targets,top_n_regulators,weighted_networks)
}
}
