% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model_EV_cargo_prediction.R
\name{get_multi_EV_cargo_importances}
\alias{get_multi_EV_cargo_importances}
\title{Get EV_cargo importances from a multi-EV_cargo classfication model.}
\usage{
get_multi_EV_cargo_importances(setting,EV_cargo_target_matrix, EV_cargo_position = "cols", algorithm, cv = TRUE, cv_number = 4, cv_repeats = 2, parallel = FALSE, n_cores = 4, ignore_errors = FALSE, continuous = TRUE, known = TRUE, filter_genes = FALSE)
}
\arguments{
\item{setting}{A list containing the following elements: .$name: name of the setting; .$from: name(s) of the EV_cargo(s) of which the predictve performance need to be assessed; .$response: the observed target response: indicate for a gene whether it was a target or not in the setting of interest. $EV_cargo: NULL or the name of the EV_cargo(s) that are known to be active in the setting of interest.}

\item{EV_cargo_target_matrix}{A matrix of EV_cargo-target probabilty scores (recommended) or discrete target assignments (not-recommended).}

\item{EV_cargo_position}{Indicate whether the EV_cargo in the EV_cargo-target matrix are in the rows ("rows") or columns ("cols"). Default: "cols"}

\item{algorithm}{The name of the classification algorithm to be applied. Should be supported by the caret package. Examples of algorithms we recommend: with embedded feature selection: "rf","glm","fda","glmnet","sdwd","gam","glmboost"; without: "lda","naive_bayes","pls"(because bug in current version of pls package), "pcaNNet". Please notice that not all these algorithms work when the features (i.e. EV_cargo vectors) are categorical (i.e. discrete class assignments).}

\item{cv}{Indicate whether model training and hyperparameter optimization should be done via cross-validation. Default: TRUE. FALSE might be useful for applications only requiring variable importance, or when final model is not expected to be extremely overfit.}

\item{cv_number}{The number of folds for the cross-validation scheme: Default: 4; only relevant when cv == TRUE.}

\item{cv_repeats}{The number of repeats during cross-validation. Default: 2; only relevant when cv == TRUE.}

\item{parallel}{Indiciate whether the model training will occur parallelized. Default: FALSE. TRUE only possible for non-windows OS.}

\item{n_cores}{The number of cores used for parallelized model training via cross-validation. Default: 4. Only relevant on non-windows OS.}

\item{ignore_errors}{Indiciate whether errors during model training by caret should be ignored such that another model training try will be initiated until model is trained without raising errors. Default: FALSE.}

\item{continuous}{Indicate whether during training of the model, model training and evaluation should be done on class probabilities or discrete class labels. For huge class imbalance, we recommend setting this value to TRUE. Default: TRUE.}

\item{known}{Indicate whether the true active EV_cargo for a particular dataset is known or not. Default: TRUE. The true EV_cargo will be extracted from the $EV_cargo slot of the setting.}

\item{filter_genes}{Indicate whether 50 per cent of the genes that are the least variable in EV_cargo-target scores should be removed in order to reduce the training of the model. Default: FALSE.}
}
\value{
A data.frame with for each EV_cargo - data set combination, feature importance scores indicating how important the query EV_cargo is for the prediction of the response in the particular dataset, when prediction is done via a trained classification model with all possible EV_cargo as input. In addition to the importance score(s), the name of the particular setting ($setting), the name of the query EV_cargo($test_EV_cargo), the name of the true active EV_cargo (if known: $EV_cargo).
}
\description{
\code{get_multi_EV_cargo_importances} A classificiation algorithm chosen by the user is trained to construct one model based on the target gene predictions of all EV_cargo of interest (EV_cargo are considered as features) in order to predict the observed response in a particular dataset. Variable importance scores that indicate for each EV_cargo the importance for response prediction, are extracted. It can be assumed that EV_cargo with higher variable importance scores are more likely to be a true active EV_cargo.
}
\examples{
\dontrun{
settings = lapply(expression_settings_validation[1:5],convert_expression_settings_evaluation)
settings_EV_cargo_pred = convert_settings_EV_cargo_prediction(settings, all_EV_cargo = unlist(extract_EV_cargo_from_settings(settings,combination = FALSE)), validation = TRUE, single = FALSE)

weighted_networks = construct_weighted_networks(lr_network, sig_network, gr_network, source_weights_df)
EV_cargo = extract_EV_cargo_from_settings(settings_EV_cargo_pred,combination = FALSE)
EV_cargo_target_matrix = construct_EV_cargo_target_matrix(weighted_networks, EV_cargo)
EV_cargo_importances_glm = dplyr::bind_rows(lapply(settings_EV_cargo_pred, get_multi_EV_cargo_importances,EV_cargo_target_matrix, algorithm = "glm"))
print(head(EV_cargo_importances_glm))
}
}
