% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model_EV_cargo_prediction.R
\name{get_single_EV_cargo_importances}
\alias{get_single_EV_cargo_importances}
\title{Get EV_cargo importances based on target gene prediction performance of single EV_cargo.}
\usage{
get_single_EV_cargo_importances(setting,EV_cargo_target_matrix, EV_cargo_position = "cols", known = TRUE)
}
\arguments{
\item{setting}{A list containing the following elements: .$name: name of the setting; .$from: name(s) of the EV_cargo(s) of which the predictve performance need to be assessed; .$response: the observed target response: indicate for a gene whether it was a target or not in the setting of interest. $EV_cargo: NULL or the name of the EV_cargo(s) that are known to be active in the setting of interest.}

\item{EV_cargo_target_matrix}{A matrix of EV_cargo-target probabilty scores (or discrete target assignments).}

\item{EV_cargo_position}{Indicate whether the EV_cargo in the EV_cargo-target matrix are in the rows ("rows") or columns ("cols"). Default: "cols"}

\item{known}{Indicate whether the true active EV_cargo for a particular dataset is known or not. Default: TRUE. The true EV_cargo will be extracted from the $EV_cargo slot of the setting.}
}
\value{
A data.frame with for each EV_cargo - data set combination, classification evaluation metrics indicating how well the query EV_cargo predicts the response in the particular dataset. Evaluation metrics are the same as in \code{\link{evaluate_target_prediction}}. In addition to the metrics, the name of the particular setting ($setting), the name of the query EV_cargo($test_EV_cargo), the name of the true active EV_cargo (if known: $EV_cargo).
}
\description{
\code{get_single_EV_cargo_importances} Get EV_cargo importance measures for EV_cargo based on how well a single, individual, EV_cargo can predict an observed response. Assess how well every EV_cargo of interest is able to predict the observed transcriptional response in a particular dataset, according to the EV_cargo-target model. It can be assumed that the EV_cargo that best predicts the observed response, is more likely to be the true EV_cargo.
}
\examples{
\dontrun{
settings = lapply(expression_settings_validation[1:5],convert_expression_settings_evaluation)
settings_EV_cargo_pred = convert_settings_EV_cargo_prediction(settings, all_EV_cargo = unlist(extract_EV_cargo_from_settings(settings,combination = FALSE)), validation = TRUE, single = TRUE)

weighted_networks = construct_weighted_networks(lr_network, sig_network, gr_network, source_weights_df)
EV_cargo = extract_EV_cargo_from_settings(settings_EV_cargo_pred,combination = FALSE)
EV_cargo_target_matrix = construct_EV_cargo_target_matrix(weighted_networks, EV_cargo)
EV_cargo_importances = dplyr::bind_rows(lapply(settings_EV_cargo_pred,get_single_EV_cargo_importances,EV_cargo_target_matrix))
print(head(EV_cargo_importances))
}
}
