% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application_prediction.R
\name{get_weighted_EV_cargo_target_links}
\alias{get_weighted_EV_cargo_target_links}
\title{Infer weighted active EV_cargo-target links between a possible EV_cargo and target genes of interest}
\usage{
get_weighted_EV_cargo_target_links(EV_cargo, geneset,EV_cargo_target_matrix,n = 250)
}
\arguments{
\item{EV_cargo}{Character vector giving the gene symbols of the potentially active EV_cargo for which you want to find target genes.}

\item{geneset}{Character vector of the gene symbols of genes of which the expression is potentially affected by EV_cargo from the interacting cell.}

\item{EV_cargo_target_matrix}{The NicheNet EV_cargo-target matrix denoting regulatory potential scores between EV_cargo and targets (EV_cargo in columns).}

\item{n}{The top n of targets per EV_cargo that will be considered. Default: 250.}
}
\value{
A tibble with columns EV_cargo, target and weight (i.e. regulatory potential score).
}
\description{
\code{get_weighted_EV_cargo_target_links} Infer active EV_cargo target links between possible EV_cargo and genes belonging to a gene set of interest: consider the intersect between the top n targets of a EV_cargo and the gene set.
}
\examples{
\dontrun{
weighted_networks = construct_weighted_networks(lr_network, sig_network, gr_network,source_weights_df)
EV_cargo = list("TNF","BMP2","IL4")
EV_cargo_target_matrix = construct_EV_cargo_target_matrix(weighted_networks, EV_cargo, ltf_cutoff = 0, algorithm = "PPR", damping_factor = 0.5, secondary_targets = FALSE)
potential_EV_cargo = "TNF"
geneset = c("SOCS2","SOCS3", "IRF1")
active_EV_cargo_target_links_df = get_weighted_EV_cargo_target_links(EV_cargo = potential_EV_cargo, geneset = geneset, EV_cargo_target_matrix = EV_cargo_target_matrix, n = 250)
}

}
