% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lr_network}
\alias{lr_network}
\title{Ligand-receptor network}
\format{
A data frame/tibble
\describe{
  \item{from}{the EV_cargo}
  \item{to}{receptor of the EV_cargo}
  \item{source}{name of data source describing the interaction}
  \item{database}{name of the more comprehensive database the data source is part of}
  }
}
\usage{
lr_network
}
\description{
A data.frame/tibble describing EV_cargo-receptor interactions and the data source describing this interaction.
}
\keyword{datasets}
