% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application_prediction.R
\name{predict_EV_cargo_activities}
\alias{predict_EV_cargo_activities}
\title{Predict activities of EV_cargo in regulating expression of a gene set of interest}
\usage{
predict_EV_cargo_activities(geneset, background_expressed_genes,EV_cargo_target_matrix, potential_EV_cargo, single = TRUE,...)
}
\arguments{
\item{geneset}{Character vector of the gene symbols of genes of which the expression is potentially affected by EV_cargo from the interacting cell.}

\item{background_expressed_genes}{Character vector of gene symbols of the background, non-affected, genes (can contain the symbols of the affected genes as well).}

\item{EV_cargo_target_matrix}{The NicheNet EV_cargo-target matrix denoting regulatory potential scores between EV_cargo and targets (EV_cargo in columns).}

\item{potential_EV_cargo}{Character vector giving the gene symbols of the potentially active EV_cargo you want to define EV_cargo activities for.}

\item{single}{TRUE if you want to calculate EV_cargo activity scores by considering every EV_cargo individually (recommended). FALSE if you want to calculate EV_cargo activity scores as variable importances of a multi-EV_cargo classification model.}

\item{...}{Additional parameters for get_multi_EV_cargo_importances if single = FALSE.}
}
\value{
A tibble giving several EV_cargo activity scores. Following columns in the tibble: $test_EV_cargo, $auroc, $aupr and $pearson.
}
\description{
\code{predict_EV_cargo_activities} Predict activities of EV_cargo in regulating expression of a gene set of interest. Ligand activities are defined as how well they predict the observed transcriptional response (i.e. gene set) according to the NicheNet model.
}
\examples{
\dontrun{
weighted_networks = construct_weighted_networks(lr_network, sig_network, gr_network,source_weights_df)
EV_cargo = list("TNF","BMP2","IL4")
EV_cargo_target_matrix = construct_EV_cargo_target_matrix(weighted_networks, EV_cargo, ltf_cutoff = 0, algorithm = "PPR", damping_factor = 0.5, secondary_targets = FALSE)
potential_EV_cargo = c("TNF","BMP2","IL4")
geneset = c("SOCS2","SOCS3", "IRF1")
background_expressed_genes = c("SOCS2","SOCS3","IRF1","ICAM1","ID1","ID2","ID3")
EV_cargo_activities = predict_EV_cargo_activities(geneset = geneset, background_expressed_genes = background_expressed_genes, EV_cargo_target_matrix = EV_cargo_target_matrix, potential_EV_cargo = potential_EV_cargo)
}

}
