% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application_prediction.R
\name{prepare_EV_cargo_target_visualization}
\alias{prepare_EV_cargo_target_visualization}
\title{Prepare heatmap visualization of the EV_cargo-target links starting from a EV_cargo-target tibble.}
\usage{
prepare_EV_cargo_target_visualization(EV_cargo_target_df, EV_cargo_target_matrix, cutoff = 0.25)
}
\arguments{
\item{EV_cargo_target_df}{Tibble with columns 'EV_cargo', 'target' and 'weight' to indicate EV_cargo-target regulatory potential scores of interest.}

\item{EV_cargo_target_matrix}{The NicheNet EV_cargo-target matrix denoting regulatory potential scores between EV_cargo and targets (EV_cargo in columns).}

\item{cutoff}{Quantile cutoff on the EV_cargo-target scores of the input weighted EV_cargo-target network. Scores under this cutoff will be set to 0.}
}
\value{
A matrix giving the EV_cargo-target regulatory potential scores between EV_cargo of interest and their targets genes part of the gene set of interest.
}
\description{
\code{prepare_EV_cargo_target_visualization} Prepare heatmap visualization of the EV_cargo-target links starting from a EV_cargo-target tibble. Get regulatory potential scores between all pairs of EV_cargo and targets documented in this tibble. For better visualization, we propose to define a quantile cutoff on the EV_cargo-target scores.
}
\examples{
\dontrun{
weighted_networks = construct_weighted_networks(lr_network, sig_network, gr_network,source_weights_df)
EV_cargo = list("TNF","BMP2","IL4")
EV_cargo_target_matrix = construct_EV_cargo_target_matrix(weighted_networks, EV_cargo, ltf_cutoff = 0, algorithm = "PPR", damping_factor = 0.5, secondary_targets = FALSE)
geneset = c("SOCS2","SOCS3", "IRF1")
background_expressed_genes = c("SOCS2","SOCS3","IRF1","ICAM1","ID1","ID2","ID3")
active_EV_cargo_target_links_df = potential_EV_cargo \%>\% lapply(get_weighted_EV_cargo_target_links, geneset = geneset, EV_cargo_target_matrix = EV_cargo_target_matrix, n = 250) \%>\% bind_rows()
active_EV_cargo_target_links = prepare_EV_cargo_target_visualization(EV_cargo_target_df = active_EV_cargo_target_links_df, EV_cargo_target_matrix = EV_cargo_target_matrix, cutoff = 0.25)
}

}
